상품 list page 및 상품 detail page 리팩토링 , 구현 마무리 상품 등록 페이지 로직 수정
